Update README.md
The purpose of this dashboard is to help an organization to improve employee performance and employee retention (reduce attrition). I’ve sourced the data from the internet. My main source of data is a .CSV file containing 38 columns and 1000+ rows.
I’ve done the data cleaning my removing the duplicate values and the null values. Later, I’ve created multiple visualizations which help the viewer to understand the factors that affect the attrition.
Here are the visuals that I’ve created:
•	Count of Overall Employees
•	Attrition
•	Attrition Rate
•	Average Age
•	Average Salary
•	Average Years
•	Attrition by Gender
•	Attrition by Education
•	Attrition by Age
•	Attrition by Salary
•	Attrition by YearsAtCompany
•	JobSatisfaction